pyproject.toml
[project]
name = "a2a-langgraph-agent-project"
version = "0.1.0"
description = "A2A, MCP, LangGraph Agent Integration"
authors = [
    {name = "Hyunjun Jeon", email = "jeonhj920@gmail.com"},
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[grpc]>=0.3.0",
    "aiohttp>=3.9.0",
    "fastmcp>=2.11.3",
    "finance-datareader>=0.9.96",
    "fredapi>=0.5.2",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.1",
    "langchain>=0.3.27",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.30",
    "langgraph>=0.6.4",
    "pandas>=2.3.1",
    "psutil>=7.0.0",
    "publicdatareader>=1.1.0",
    "python-dotenv>=1.1.1",
    "redis>=5.0.0",
    "structlog>=25.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0,<8.0.0",
    "pytest-asyncio>=0.21.0,<0.22.0",
    "ruff>=0.1.0,<0.2.0",
    "playwright>=1.48.0",
    "aiohttp>=3.9.0",
]
test = [
    "pytest>=7.0.0,<8.0.0",
    "pytest-asyncio>=0.21.0,<0.22.0",
    "coverage>=7.0.0,<8.0.0",
]


[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "strict"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"
include = ["src/**/*.py"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
extend-ignore = [
    "E501",  # Line too long (handled by black)
    "C901",  # Complex structure (delegated to design)
    "E402",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*"]
